+++ src/api/index.ts
@@ -2,3 +2,4 @@ export * from './api';
 export * from './request';
 export * from './search';
 export * from './browse';
+export * from './preview';